CASMINST-3919: Set podAntiAffinity on the pooler
The pooler pods could wind up all running on the same node so become
unavailable if a node goes down.

This adds podAntiAffinity to the pooler pods so that they will not
run on the same node.

I found that when there are 3 nodes (like on TDS systems) there
isn't room to start the new pod when rolling the deployment. This
would cause problems since we couldn't roll out a fix. The
solution to this is to set the .spec.strategy.rollingUpdate.maxSurge
to 0. With this setting one of the old replicas is terminated
along with the new replica that's starting.